# pkg/client/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  title: Feed API
  version: "1.0"
servers:
- url: //localhost:8080/v1
paths:
  /articles/{id}:
    get:
      description: view article
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.ArticleResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: view article
  /feeds:
    get:
      description: list feeds
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.FeedsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: list feeds
    post:
      description: create feed
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/server.CreateFeedRequest'
        description: feed data
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.CreateResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: create feed
      x-codegen-request-body-name: request
  /feeds/{id}:
    get:
      description: view feed
      parameters:
      - description: Feed ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.FeedResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.ErrorResponse'
          description: Internal Server Error
      summary: view feed
    put:
      description: update feed
      parameters:
      - description: Feed ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/server.UpdateFeedRequest'
        description: feed data
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.UpdateResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: update feed
      x-codegen-request-body-name: request
  /feeds/{id}/articles:
    get:
      description: list articles for a feed
      parameters:
      - description: Feed ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.FeedArticlesResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: list articles for a feed
components:
  schemas:
    server.ArticleResponse:
      example:
        article:
          preview: preview
          authorEmail: authorEmail
          feedId: feedId
          authorName: authorName
          imageUrl: imageUrl
          guid: guid
          id: id
          title: title
          content: content
          url: url
      properties:
        article:
          $ref: '#/components/schemas/service.Article'
      type: object
    server.CreateFeed:
      properties:
        name:
          example: My Feed
          type: string
        url:
          example: https://example.com/feed.xml
          type: string
      required:
      - url
      type: object
    server.CreateFeedRequest:
      properties:
        feed:
          $ref: '#/components/schemas/server.CreateFeed'
      type: object
    server.CreateResponse:
      example:
        id: id
      properties:
        id:
          type: string
      type: object
    server.ErrorResponse:
      example:
        error: unable to fetch feeds
      properties:
        error:
          example: unable to fetch feeds
          type: string
      type: object
    server.FeedArticlesResponse:
      example:
        articles:
        - preview: preview
          authorEmail: authorEmail
          feedId: feedId
          authorName: authorName
          imageUrl: imageUrl
          guid: guid
          id: id
          title: title
          content: content
          url: url
        - preview: preview
          authorEmail: authorEmail
          feedId: feedId
          authorName: authorName
          imageUrl: imageUrl
          guid: guid
          id: id
          title: title
          content: content
          url: url
      properties:
        articles:
          items:
            $ref: '#/components/schemas/service.Article'
          type: array
      type: object
    server.FeedResponse:
      example:
        feed:
          name: name
          id: id
          url: url
      properties:
        feed:
          $ref: '#/components/schemas/service.Feed'
      type: object
    server.FeedsResponse:
      example:
        feeds:
        - name: Example
          id: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
          url: https://example.com/
        - name: Example
          id: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
          url: https://example.com/
      properties:
        feeds:
          items:
            $ref: '#/components/schemas/server.ListFeed'
          type: array
      type: object
    server.ListFeed:
      example:
        name: Example
        id: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
        url: https://example.com/
      properties:
        id:
          example: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
          type: string
        name:
          example: Example
          type: string
        url:
          example: https://example.com/
          type: string
      type: object
    server.UpdateFeed:
      properties:
        name:
          example: My Feed
          type: string
        url:
          example: https://example.com/feed.xml
          type: string
      required:
      - url
      type: object
    server.UpdateFeedRequest:
      properties:
        feed:
          $ref: '#/components/schemas/server.UpdateFeed'
      type: object
    server.UpdateResponse:
      example:
        id: id
      properties:
        id:
          type: string
      type: object
    service.Article:
      example:
        preview: preview
        authorEmail: authorEmail
        feedId: feedId
        authorName: authorName
        imageUrl: imageUrl
        guid: guid
        id: id
        title: title
        content: content
        url: url
      properties:
        authorEmail:
          type: string
        authorName:
          type: string
        content:
          type: string
        feedId:
          type: string
        guid:
          type: string
        id:
          type: string
        imageUrl:
          type: string
        preview:
          type: string
        title:
          type: string
        url:
          type: string
      type: object
    service.Feed:
      example:
        name: name
        id: id
        url: url
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
      type: object
x-original-swagger-version: "2.0"